Change product weight type to Double

+++ corpis-domain-jpa/corpis-product-domain-jpa/src/main/java/sk/seges/corpis/server/domain/product/jpa/JpaProduct.java
@@ -120,7 +120,7 @@ public class JpaProduct extends ProductBase {
 
 	@Override
 	@Column(name = "weight")
-	public Float getWeight() {
+	public Double getWeight() {
 		return super.getWeight();
 	}
