+++ corpis-domain/corpis-product-domain-model/src/main/java/sk/seges/corpis/server/domain/product/Product.java
@@ -1,5 +1,9 @@
 package sk.seges.corpis.server.domain.product;
 
+import java.util.Date;
+import java.util.List;
+import java.util.Set;
+
 import sk.seges.corpis.appscaffold.shared.annotation.BaseObject;
 import sk.seges.corpis.appscaffold.shared.annotation.DomainInterface;
 import sk.seges.corpis.server.domain.Description;
@@ -11,10 +15,6 @@ import sk.seges.sesam.domain.IMutableDomainObject;
 import sk.seges.sesam.pap.model.annotation.ReadOnly;
 import sk.seges.sesam.security.shared.domain.ISecuredObject;
 
-import java.util.Date;
-import java.util.List;
-import java.util.Set;
-
 @DomainInterface
 @BaseObject
 public interface Product extends IMutableDomainObject<Long>, ISecuredObject<Long> {
@@ -24,7 +24,7 @@ public interface Product extends IMutableDomainObject<Long>, ISecuredObject<Long
 	List<? extends Name> names();
 	List<? extends Description> descriptions();
 	String description();
-	Float weight();
+	Double weight();
 	Integer unitsPerPackage();
 	List<? extends ProductPrice> prices();
 	Set<? extends ProductPrice> fees();
